requirements.txt
python-binance
pandas
numpy
ta
python-dotenv
openai
httpx
